# perf: don't access array out of bounds in pull.values