@dominictarr 2016-09-04
okay cool. pull.values is mainly for testing, but will merge anyway. thanks!