@dominictarr 2016-09-04
merged into 3.4.5